Adding ObsidianVault
This projects was in the works about two weeks prior to officially publishing it and getting my usual setup working.  In uploading it, I'm saving my documentation in my usual Obsidian Vault.
As well, gamepads now use the D-Pad to control the player character.

This is to be my main project - I'm going to give it my all!
:camel: