DOC & TEST - Approaching Special Tiles
Like the tiles in Metroid and other 2d games, I'm looking to try to make some common behavior for tiles that gets more specific depending on the type. Levers will have to trigger SOMETHING or some event, and doors have to take you SOMEWHERE or to some location. Most of my study is in my notes.

:camel: