DEV - Implemented basic BreakableTile
With some of the work I've done so far, I've successfully implemented a Breakable tile like in Metroid, that works via events very well so far! I'm so glad!

I just have to see about the rules for regenerating if the player or some other object is within the block...

:camel: