DEV - Destroy Projectiles within BreakableTile
I violated the Single Responsibility Principle... I'm a bad man....
I did fix it though, by ensuring that the Projectiles don't know about the Tile besides not immediately destroying themselves against it. In the BreakableTile class, I just have it so that they destroy the Projectiles if we don't allow them through the Tile.
Tested, and it still works fine thus far.

:camel: