DEV - Working more on InteractableTile
After I get this done, I need to make the prefab for it and evaluate this iteration...
I'm doing a questionable thing with whether the InteractableTile is a toggle or not... It stinks of redundancy but saves me from making another event.
I also have to think more about how I'll implement the effects/responses to that event... I'm so excited!

:camel: