DEV & ART - Hashing out Leveltile Workflow
I've drawn generic icons that will be hidden during runtime for my LevelTiles, so that all animated and level-specific sprite work is to be handling within the LevelTileGO on that tile. This helps me as I edit the maps and allows me to animate the tiles in a more sensical workflow as I develop this solution.

:camel: