DOC - Moving InteractableTiles "off-Tilemap"
After a whole lot of work last night, I'm deciding to, for ease of workflow, pre-define my InteractableTiles in the editor AND their connections. This seems to make the most sense, I think, since the very anonymous-prefab workflow I was going with before required me to externally (and unintuitively) define all the connections and things like that that I'd do. With this system, I should be able to make my InteractableTiles hook up to things more easily.

:camel: