ART - Drew Max
I drew a little sprite of my dog, Max. He's a Golden Retriever, which is apparently a Scottish breed. I looked up what Scots wore around the 15th century but didn't get a lot of results, so I gave him a plaid-esque tunic and saffron shirt, bare-legged look with a regional poncho on top that was given as a gift. Maybe he can fit somewhere in the game, lol!

:camel: :dog2: