DEV - Migrating code to PlayerMoveState
In my approach to turn PlayerController into a FSM context class, I had to roughly migrate some of my movement code into a new class. NOTE THIS IS NOT MY FINAL ANSWER THIS SUCKS AND I HAVE TO DECOUPLE IT MORE!!

We have to still make a Jump state and/or a way to transition into an InAir state or something like that.

:camel: