DOC + DEV - Detecting Directional Skills
I found out how to listen for direction-oriented skill inputs which was very  easy. However, finding out what to do in reaction to that was more difficult as it related to the capabilities of different characters. In this case, I think it a forgivable offense (if at all) to create two PlayerControllers for Achik and Kowi that will have their own states as to how they control and things like that. I'll try breaking up some of these things in the next commits.

:camel: