MUSIC - Wrote a song I liked
I wrote this song for the little village that you spawn in currently. The name Achpakatambo translates roughly to "farmer's village/settlement" which is the vibe I tried to evoke with this tune and its gentle nature. I forgave myself in using the charango sound, despite the historical inaccuracy; the charango was invented post-Spanish conquest. In any case,  I drew the line as to how hard I wish to stick to history and how I can better represent the Inca and Andean peoples with their instruments in song.

In short, I'm doing it anyway, and I like the initial results.

:camel: :musical_keyboard: