DOC - Player Files, Attiniy, and more
I started by returning from vacation and reorganizing my project so that I have a new folder in my Obsidian vault about the World and its Characters, specifically Achik. There, I had the idea for a magic resource meter for Achik, called Attiniy, which will help me limit and put a leash on Achik's capabilities as he navigates the map and all that.

:camel: :sun_with_face: