DEV - Implementing SkillStates & Transition Logic
Nobody wants to start a skill and then have it interrupted by another skill activating that's bound to another Interaction (the problem). Thus, I created a sufficient abstraction and logic to find out if a skill state is done (by checking whether its component is done yet) and changing OR NOT changing state based on that.

:camel: